requirements.txt
gspread==6.2.1
oauth2client==4.1.3
playwright==1.58.0
python-dotenv==1.2.2
